# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAER301 API Documentation",
        "description": "Auto-generated API documentation for SAE301 project",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Production API Server"
        }
    ],
    "paths": {
...
        "/api/me/managed-races": {
            "get": {
                "tags": [
                    "Race Management"
                ],
                "summary": "Get races managed by the authenticated user",
                "description": "Returns a list of races where the user is an organizer, raid manager, or admin",
                "operationId": "getManagedRaces",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Race"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/races/{race}/participants": {
            "get": {
                "tags": [
                    "Race Management"
                ],
                "summary": "Get participants for a specific race",
                "description": "Returns a detailed list of participants and their registration status for a race managed by the user",
                "operationId": "getRaceParticipants",
                "parameters": [
                    {
                        "name": "race",
                        "in": "path",
                        "description": "ID of the race",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "registration_id": {
                                                "type": "integer"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "first_name": {
                                                "type": "string"
                                            },
                                            "last_name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "team_name": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "is_leader": {
                                                "type": "boolean"
                                            },
                                            "has_license": {
                                                "type": "boolean"
                                            },
                                            "has_pps": {
                                                "type": "boolean"
                                            },
                                            "docs_validated": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - User does not manage this race"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/registrations/{registration}/validate-docs": {
            "patch": {
                "tags": [
                    "Race Management"
                ],
                "summary": "Validate or flag documents for a registration",
                "description": "Allows a manager to manually validate registration documents or flag them as invalid",
                "operationId": "validateRegistrationDocs",
                "parameters": [
                    {
                        "name": "registration",
                        "in": "path",
                        "description": "ID of the registration",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "confirmed",
                                            "missing_credentials"
                                        ],
                                        "example": "confirmed"
                                    },
                                    "admin_notes": {
                                        "type": "string",
                                        "example": "Document illisible, merci de renvoyer."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Race": {
                "title": "Race",
                "description": "Orienteering race model",
                "properties": {
                    "race_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "race_name": {
                        "type": "string",
                        "example": "La Boussole de la Forêt"
                    },
                    "race_date_start": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-10-12T09:00:00Z"
                    },
                    "race_date_end": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-10-12T17:00:00Z"
                    },
                    "race_reduction": {
                        "type": "number",
                        "format": "float",
                        "example": 5,
                        "nullable": true
                    },
                    "race_meal_price": {
                        "type": "number",
                        "format": "float",
                        "example": 8,
                        "nullable": true
                    },
                    "race_duration_minutes": {
                        "type": "number",
                        "format": "float",
                        "example": 150,
                        "nullable": true
                    },
                    "raid_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "adh_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "pac_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "pae_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "dif_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "typ_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}